audit_apps/app/requirements.txt
numpy
shazamio
librosa
numpy
matplotlib
scikit-learn
requests
tinytag
mysql-connector-python
yt_dlp
discogs_client
